Added implementation file for LTexture class
